feat: implement movement simulation logic and refactor minmax functions for better game state evaluation
